fixed logis issues

+++ business_logic/Services/LocationService.cs
@@ -8,11 +8,13 @@ namespace business_logic.Services
     public class LocationService : ILocationService
     {
         private readonly IRepository<Location> _repo;
+        private readonly IRepository<Route> _routeRepo; 
         private readonly IMapper _mapper;
 
-        public LocationService(IRepository<Location> repo, IMapper mapper)
+        public LocationService(IRepository<Location> repo, IRepository<Route> routeRepo, IMapper mapper)
         {
             _repo = repo;
+            _routeRepo = routeRepo; 
             _mapper = mapper;
         }
 
@@ -26,14 +28,20 @@ namespace business_logic.Services
         {
             var loc = _repo.GetById(id);
             if (loc == null) throw new Exception("Location not found");
+
+            var isUsedInRoutes = _routeRepo.GetAll()
+                .Any(r => r.StartLocationId == id || r.DestinationLocationId == id);
+
+            if (isUsedInRoutes)
+            {
+                throw new InvalidOperationException("Unable to delete location");
+            }
+
             _repo.Delete(id);
             _repo.Save();
         }
 
-        public IEnumerable<LocationDTO> GetAll()
-        {
-            return _mapper.Map<IEnumerable<LocationDTO>>(_repo.GetAll());
-        }
+        public IEnumerable<LocationDTO> GetAll() => _mapper.Map<IEnumerable<LocationDTO>>(_repo.GetAll());
 
         public async Task<LocationDTO> Get(int id)
         {
